+++ CHANGELOG.md
@@ -14,10 +14,14 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Changes
 - Renamed `utils.lua` to `ari_utils.lua` to prevent interference with other lua plugins
+- Added subtrees "Extra Information" and "Unknown Bits" to TLV tree to clear up the overview
 
 ### Fixes
 - Remove unknown bit 7 in byte 5 (is part of the group id) by @sterzy (#4)
 - Fix range of bytes shown to include parts of the group id (#8)
+
+### Removals
+- Removed TLV "Starts at" field
 
 ## [v1.0.0](https://github.com/seemoo-lab/aristoteles/tree/v1.0.0) (11-08-2021)